Port over SLURM backend
