+++ continuous_integration/docker/slurm/start.sh
@@ -9,4 +9,7 @@ docker run --rm -d \
     --name slurm \
     -h slurm \
     -v "$git_root":/working \
+    -p 8000:8000 \
+    -p 8786:8786 \
+    -p 8088:8088 \
     daskgateway/testing-slurm
